king: eyre: Hooked up multi-tenet HTTP. Start-up is a bit hacky. Still totally untested.
